// resources/js/types/About.ts
export interface AboutProps {
    id: number;
    title: string;
    about: string;
    history: string;
    achivement: string;
    purpose: string;
    image: string;
    logo: string;
    logo2: string;
    vision: string;
    mission: string;
    phone: string;
    email: string;
    address: string;
    facebook: string;
    twitter: string;
    instagram: string;
    linkedin: string;
    created_at: string;
    updated_at: string;
}